% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs_analysis.R
\name{analyse_meta}
\alias{analyse_meta}
\title{Analyse the robust regression results by providing a table with descriptives}
\usage{
analyse_meta(coeff, t_values)
}
\arguments{
\item{coeff}{A dataframe of coefficients}

\item{t_values}{A dataframe of t-values}
}
\value{
A dataframe with the descriptives
}
\description{
Analyse the robust regression results by providing a table with descriptives
}
\examples{
result <- rrs(mpg ~., data = mtcars, k = 4)
coeff <- result[1]
t_values <- result[2]
analyse_meta(coeff, t_values)
}
