% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrs_bp.R
\name{rrs_bp}
\alias{rrs_bp}
\title{Creates a barplot of the rrs results}
\usage{
rrs_bp(analysis, rm_constant = FALSE)
}
\arguments{
\item{analysis}{a dataframe resulting from the \code{\link{rrs_analysis}} function}

\item{rm_constant}{If there is a constant in the `analysis` dataframe, remove this by setting this at `TRUE`, default is `FALSE`}
}
\value{
a ggplot object
}
\description{
Creates a barplot of the rrs results
}
\examples{
result <- rrs(mpg ~., data = mtcars, k = 4)
coeff <- result[1]
t_values <- result[2]
analysis <- rrs_analysis(coeff, t_values)
rrs_bp(analysis)
}
