% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrsim.R
\docType{package}
\name{rrsim}
\alias{rrsim}
\alias{rrsim-package}
\title{rrsim: a package for a rubust regression simulation}
\description{
The rrism package provides two important functions:
rrs and analyse_meta.
}
\section{rrsim functions}{

The rrsim functions
}

